requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.7
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-environ==0.11.2
django-js-asset==2.2.0
environs==11.0.0
marshmallow==3.21.3
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.2
tzdata==2021.5
